rename files to aproptiate file structure
